requirements.txt
flask 
flask_sqlalchemy 
flask_jwt_extended
psycopg2-binary #PostgreSQL
python-dotenv
